Update send_wait_time text to reflect modbus timing changes (#4399)
Documentation update to reflect change in how send_wait_time is handled in modbus.

Previously send_wait_time had to be set to the longest possible message on the bus, which is difficult to calculate, and error prone. It would clobber incoming messages by sending outgoing commands in the middle of receipt.

New functionality (see https://github.com/esphome/esphome/pull/7674 ) changes this to wait for the response to be complete, and then sending the next command.

Now send_wait_time only needs to be large enough to allow time for devices to send their first byte of response.